games_similar, links php error fixes

+++ Website/AtariLegend/admin/games/games_similar.php
@@ -13,13 +13,12 @@
 
 //load all common functions
 include("../includes/common.php"); 
-include("../includes/config.php"); 
 
 
 //***********************************************************************************
 //If delete similar link has been pressed
 //***********************************************************************************
-if ( $action == 'delete_similar' )
+if ( isset($action) and $action == 'delete_similar' )
 {
 	$sql_aka = mysql_query("DELETE FROM game_similar WHERE game_similar_id = '$game_similar_id' AND game_id = '$game_id'") 
 			   or die ("Couldn't delete similar game");
@@ -28,7 +27,7 @@ if ( $action == 'delete_similar' )
 //***********************************************************************************
 //If add similar button has been pressed
 //***********************************************************************************
-if ( $action == 'game_similar' )
+if ( isset($action) and $action == 'game_similar' )
 {
 	$sql_aka = mysql_query("INSERT INTO game_similar (game_id, game_similar_cross) VALUES ('$game_id','$game_similar')")
  			   or die ("Couldn't insert similar game");
@@ -46,8 +45,8 @@ if ( $action == 'game_similar' )
 		$game_info=mysql_fetch_array($sql_game);
 		
 			$smarty->assign('game_info',
-	    		 array('game_name' => $game_info[game_name],
-				 	   'game_id' => $game_info[game_id]));
+	    		 array('game_name' => $game_info['game_name'],
+				 	   'game_id' => $game_info['game_id']));
 		
 
 //***********************************************************************************
@@ -59,13 +58,13 @@ if ( $action == 'game_similar' )
 							    LEFT JOIN game ON (game_similar.game_similar_cross = game.game_id)
 								WHERE game_similar.game_id='$game_id'")
 					or die ("Couldn't query similar games");
-	
+	$nr_similar = 0;
 	while  ($similar=mysql_fetch_array($sql_similar)) 
 	{ 
 		$smarty->append('similar',
-	   		 	 array('game_similar_id' => $similar[game_similar_id],
-				 	   'game_name' => $similar[game_name],
-					   'game_id' => $similar[game_similar_cross]));
+	   		 	 array('game_similar_id' => $similar['game_similar_id'],
+				 	   'game_name' => $similar['game_name'],
+					   'game_id' => $similar['game_similar_cross']));
 		$nr_similar++;
 	}
 	
@@ -77,13 +76,13 @@ if ( $action == 'game_similar' )
 	while  ($game_similar=mysql_fetch_array($sql_game_similar)) 
 	{ 
 		$smarty->append('similar_games',
-	   		 	 array('game_id' => $game_similar[game_id],
-				 	   'game_name' => $game_similar[game_name]));
+	   		 	 array('game_id' => $game_similar['game_id'],
+				 	   'game_name' => $game_similar['game_name']));
 	}
 
 
 
-		$smarty->assign("user_id",$_SESSION[user_id]);
+		$smarty->assign("user_id",$_SESSION['user_id']);
 		$smarty->assign('games_similar_tpl', '1');
 
 		//Send all smarty variables to the templates
